# perf(lex): use linear search over regex