Resolves #513.

For the following table:
- Cppfront is built with the respective Cpp1 compiler configuration.
- LLVM is shorthand for Clang and Libc++.

Legend:
- `main`: The regex is an automatic object (branch `main`).
- `find_if`: The keywords are stored in `static` vector and queried with `std::find_if`.

| `mixed-as-for-variant-20-types` | LLVM 17, Debug | LLVM 17, Release | GCC 13, Debug | GCC 13, Release |
|:--------------------------------|:---------------|:-----------------|:--------------|:----------------|
| Cppfront time      (`main`)     | 11.28 s        | 0.94 s           | 3.44 s        | 0.45 s          |
| Cppfront time      (`find_if`)  | 0.08 s         | 0.01 s           | 0.08 s        | 0.01 s          |

| [#513 minimal reproducer][]    | LLVM 17, Debug | LLVM 17, Release | GCC 13, Debug | GCC 13, Release |
|:-------------------------------|:---------------|:-----------------|:--------------|:----------------|
| Cppfront time      (`main`)    | 14.30 s        | 1.06 s           | 2.53 s        | 0.34 s          |
| Cppfront time      (`find_if`) | 0.03 s         | 0.02 s           | 0.04 s        | 0.01 s          |

| `reflect.h`               | LLVM 17, Debug | LLVM 17, Release | GCC 13, Debug | GCC 13, Release |
|:--------------------------|:---------------|:-----------------|:--------------|:----------------|
| Cppfront time (`main`)    | 13.19 s        | 1.27 s           | 6.03 s        | 0.75 s          |
| Cppfront time (`find_if`) | 0.19 s         | 0.02 s           | 0.19 s        | 0.01 s          |

<a id="test"/>
<!-- <details><summary>
<a href="#test">T</a>esting summary:
</summary> -->

**[T](#est)esting summary**:

```ctest
100% tests passed, 0 tests failed out of 660

Total Test time (real) =  96.59 sec
```

<!-- </details> -->

<a id="ack"/>

**[A](#ack)cknowledgements**:

<!-- <details><summary>
<a href="#ack">A</a>cknowledgements:
</summary> -->

- Tested thanks to <https://github.com/modern-cmake/cppfront>.
- The commit message uses [Conventional Commits][] 1.0.0.

<!-- </details> -->

[Conventional Commits]: https://www.conventionalcommits.org/
[#513 minimal reproducer]: https://github.com/hsutter/cppfront/issues/513#rep